split axis plotting into separate module
